+++ src/components/commonLayout/topnav/aux/dynamic.tsx
@@ -6,7 +6,7 @@ import { DownIcon } from '@/components/icon';
 import { logout } from '@/app/actions/logout.action';
 import { Menu, MenuButton, MenuItem, MenuItems, Popover, PopoverButton, PopoverPanel } from '@headlessui/react';
 import { LogOutIcon, UserIcon } from 'lucide-react';
-import { blankButton } from '@/style/button';
+import { blankButton } from '@/styles/button';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { isTokenExists } from '@/util/auth.util';
 import Skeleton from '@/components/composite/skeleton';
