+++ src/styles/button.ts
@@ -1,7 +1,16 @@
 import { tv } from 'tailwind-variants';
 
 export const buttonBase = tv({
-  base: ['flex', 'justify-center', 'items-center', 'rounded-lg', 'transition-all', 'p-2', 'whitespace-nowrap'],
+  base: [
+    'flex',
+    'justify-center',
+    'items-center',
+    'rounded-lg',
+    'transition-all',
+    'p-2',
+    'whitespace-nowrap',
+    'active:scale-95',
+  ],
   variants: {
     disabled: {
       true: 'bg-gray-300 cursor-not-allowed opacity-50',
@@ -12,11 +21,11 @@ export const buttonBase = tv({
 
 export const filledButton = tv({
   extend: buttonBase,
-  base: ['bg-primary-500', 'dark:bg-gray-700', 'text-white', 'dark:text-white'],
+  base: 'bg-primary-500 hover:bg-primary-400 dark:bg-gray-700 text-white dark:text-white',
   variants: {
     disabled: {
       true: 'bg-gray-300 dark:bg-gray-700 border-none opacity-25',
-      false: 'bg-primary-500 cursor-pointer border-none text-white dark:text-white',
+      false: ['border-none'],
     },
   },
 });
