+++ src/styles/input.ts
